@@ -3,6 +3,7 @@
   "description": "Check and format markdown or MDX using Prettier as a remark plugin",
   "version": "1.3.0",
   "main": "dist/index.js",
+  "type": "module",
   "files": [
     "dist"
   ],
@@ -17,7 +18,7 @@
   },
   "homepage": "https://github.com/remcohaszing/remark-prettier#readme",
   "engines": {
-    "node": ">=12.0.0"
+    "node": ">=14.0.0"
   },
   "keywords": [
     "markdown",
@@ -29,33 +30,31 @@
   ],
   "scripts": {
     "prepack": "tsc --project tsconfig.build.json",
-    "test": "jest"
+    "test": "tsx src/index.test.ts"
   },
   "dependencies": {
-    "mdast-util-to-markdown": "^0.6.0",
+    "@types/mdast": "^3.0.0",
+    "@types/prettier": "^2.0.0",
+    "mdast-util-to-markdown": "^1.0.0",
     "prettier-linter-helpers": "^1.0.0",
-    "vfile-location": "^3.0.0"
+    "vfile-location": "^4.0.0"
   },
   "peerDependencies": {
     "prettier": "^2"
   },
   "devDependencies": {
-    "@types/jest": "^27.0.0",
     "@types/prettier-linter-helpers": "^1.0.0",
     "@typescript-eslint/eslint-plugin": "^5.0.0",
-    "eslint": "^7.0.0",
-    "eslint-config-remcohaszing": "^3.0.0",
-    "eslint-plugin-jest": "^25.0.0",
-    "eslint-plugin-jest-formatting": "^3.0.0",
-    "jest": "^27.0.0",
+    "eslint": "^8.0.0",
+    "eslint-config-remcohaszing": "^6.0.0",
     "memfs": "^3.0.0",
     "prettier": "^2.0.0",
-    "remark": "^13.0.0",
-    "remark-cli": "^9.0.0",
-    "ts-jest": "^27.0.0",
-    "ts-node": "^10.0.0",
+    "remark": "^14.0.0",
+    "remark-cli": "^11.0.0",
+    "tsx": "^3.8.2",
     "typescript": "^4.0.0",
-    "unified": "^9.0.0",
-    "vfile": "^4.0.0"
+    "unified": "^10.0.0",
+    "uvu": "^0.5.0",
+    "vfile": "^5.0.0"
   }
 }
